[SDL] Fix LockMousePosition handling on Unix
Because on Unix setting the cursor position is not immediate, we cannot use it
to prevent further update to CurrentMouseDelta when lock is in effect. So
instead we store the mouse position in the window coordinate when lock starts
and we simply check if the new event coordinate is the same to find out if we
should update the CurrentMouseDelta.